requirements.txt
fastapi
uvicorn
pydantic
mysql-connector-python
pytest
requests